Updated all of admin components

+++ src/app/components/admin/delete-car/delete-car.component.ts
@@ -14,7 +14,8 @@ export class DeleteCarComponent implements OnInit {
 
   public constructor(private adminService: AdminService, private router: Router) { }
 
-  ngOnInit() {
+  public ngOnInit(): void {
+   
   }
 
   public getCar():void {
@@ -29,24 +30,28 @@ export class DeleteCarComponent implements OnInit {
         this.car.image = car.image);
       this.router.navigate(["/admin/delete-car/car-id/"+this.car.id]);
     }, err => {
-      console.log(`Failed on get Car ID: `,this.car.id + `\n` +err.message);
-      alert(`Error on get Car! Wrong ID: ${this.car.id}` +` `+ `\n`+err.message);
+      console.log(`Failed on get Car! Wrong ID: `,this.car.id + `\n` +err.message);
+      alert(`Error on view Car! ` + `\n` + `The reasons: ` + `\n` + 
+      `1. No internet connection` + `\n` + 
+      `2. No connection to the server` + `\n` + 
+      `3. Wrong ID: ${this.car.id}`);
     });
   }
 
   public deleteCar(): void {
-    // confirm(`Are You sure You want to remove this Car?
-    // Id: ${this.car.id}
-    // `);
-
-    this.adminService.deleteCar(this.car.id).subscribe((c) => {
+    if(confirm(`Are You sure You want to remove this Car? ` + `\n` + `Car ID: ${this.car.id}`)) {
+      this.adminService.deleteCar(this.car.id).subscribe((c) => {
       console.log(`Success on delele Car Id: `,this.car.id = c.id);
         alert(`Car Id: ${c.id} Number: `+c.number+ ` has been succesfully deleted!`);
         this.router.navigate(["/admin/view-all-cars"]);
     }, err => {
-      console.log(`Failed on delele Car Id: `,this.car.id + `\n` +err.message);
-      alert(`Error on delete Car! Wrong Id: ${this.car.id}` +` `+ `\n`+err.message);
+      console.log(`Failed on delele Car! Wrong ID: `,this.car.id + `\n` +err.message);
+      alert(`Error on delele Car! ` + `\n` + `The reasons: ` + `\n` + 
+      `1. No internet connection` + `\n` + 
+      `2. No connection to the server` + `\n` + 
+      `3. Wrong ID: ${this.car.id}`);
     });
+   } 
   }
